Created a microservice notification which will persist the notification message in db if user is not fraud
Also created a module called clients which will store interface for the clients to the microservices. I have used feign for the same. Now using this there will be no duplicate dto class and suppose 2 micro service are there A and B, they will communicate via the clients module. Service class of A will call the interface of clients module